Day 10 impractically slow recursive solution
Loops through each light/joltage index, determines which buttons affect that index, and recursively explores all combinations. Impractically slow for the actual problem input.

Data is already stored conceptually as a matrix (implemented as a 2D array). Next step could be to implement row reducing to solve the system of linear equations or just to reduce the number of free variables.